+++ priv/repo/migrations/20250801211702_add_position_ambiguity_to_packets.exs
@@ -1,0 +1,9 @@
+defmodule Aprsme.Repo.Migrations.AddPositionAmbiguityToPackets do
+  use Ecto.Migration
+
+  def change do
+    alter table(:packets) do
+      add :position_ambiguity, :integer
+    end
+  end
+end